+++ src/main/java/org/truffleruby/parser/ast/DStrParseNode.java
@@ -39,24 +39,14 @@ import org.truffleruby.parser.ast.visitor.NodeVisitor;
 
 /** A string which contains some dynamic elements which needs to be evaluated (introduced by #). */
 public class DStrParseNode extends DParseNode implements ILiteralNode {
-    private boolean frozen;
 
-    public DStrParseNode(SourceIndexLength position, Encoding encoding, boolean frozen) {
+    public DStrParseNode(SourceIndexLength position, Encoding encoding) {
         super(position, encoding);
-        setFrozen(frozen);
     }
 
     @Override
     public NodeType getNodeType() {
         return NodeType.DSTRNODE;
-    }
-
-    public boolean isFrozen() {
-        return frozen;
-    }
-
-    public void setFrozen(boolean frozen) {
-        this.frozen = frozen;
     }
 
     /** Accept for the visitor pattern.