+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Compatibility:
 * All Array methods now return Array instances and not subclasses (#2510, @Strech).
 * Integer#zero? overrides Numeric#zero? for optimization (#2453).
 * Default `Kernel#eval` source file and line to `(eval):1` like CRuby 3 (#2453).
+* Update interpolated strings to not be frozen (#2453).
 
 Performance: